Moved /input folder to /platform/input
