+++ src/platform/input/mouse-event.js
