+++ src/platform/input/touch-device.js
@@ -1,4 +1,4 @@
-import { EventHandler } from '../core/event-handler.js';
+import { EventHandler } from '../../core/event-handler.js';
 
 import { TouchEvent } from './touch-event.js';
