+++ src/platform/input/touch-event.js
